# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "基于 Gin + Gorm + Wire 的企业级后端服务",
    "title": "Sponge API",
    "termsOfService": "https://github.com/Mojitocean/sponge",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:18888",
  "basePath": "/gateway",
  "paths": {
...
    "/gateway/follow/action": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "关注或取消关注用户",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "关注模块"
        ],
        "summary": "关注/取消关注操作",
        "parameters": [
          {
            "description": "关注信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.FollowActionReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    },
    "/gateway/follow/list": {
      "get": {
        "description": "获取用户的关注列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "关注模块"
        ],
        "summary": "获取关注列表",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "用户ID",
            "name": "user_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 30,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.UserRelationRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/follower/list": {
      "get": {
        "description": "获取用户的粉丝列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "关注模块"
        ],
        "summary": "获取粉丝列表",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "用户ID",
            "name": "user_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 30,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.UserRelationRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/friend/list": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取互相关注的好友列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "关注模块"
        ],
        "summary": "获取好友列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 30,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.UserRelationRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/like/action": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "点赞或取消点赞视频",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "点赞模块"
        ],
        "summary": "点赞/取消点赞操作",
        "parameters": [
          {
            "description": "点赞信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.LikeActionReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    },
    "/gateway/like/list": {
      "get": {
        "description": "获取用户点赞的视频列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "点赞模块"
        ],
        "summary": "获取点赞列表",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "用户ID",
            "name": "user_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 30,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.VideoInfoRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dto.FollowActionReq": {
      "type": "object",
      "required": [
        "action",
        "to_user_id"
      ],
      "properties": {
        "action": {
          "description": "1-关注, 2-取消关注",
          "type": "integer",
          "enum": [
            1,
            2
          ]
        },
        "to_user_id": {
          "type": "string",
          "example": "0"
        }
      }
    },
    "dto.LikeActionReq": {
      "type": "object",
      "required": [
        "action",
        "video_id"
      ],
      "properties": {
        "action": {
          "description": "1-点赞, 2-取消点赞",
          "type": "integer",
          "enum": [
            1,
            2
          ]
        },
        "video_id": {
          "type": "string",
          "example": "0"
        }
      }
    },
...
    "dto.UserInfoRes": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string"
        },
        "backgroundImage": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string",
          "example": "0"
        },
        "nickName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "signature": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        }
      }
    },
    "dto.UserRelationRes": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string"
        },
        "backgroundImage": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string",
          "example": "0"
        },
        "is_follow": {
          "description": "当前用户是否关注该用户",
          "type": "boolean"
        },
        "is_followed": {
          "description": "该用户是否关注当前用户",
          "type": "boolean"
        },
        "is_mutual_follow": {
          "description": "是否互关",
          "type": "boolean"
        },
        "nickName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "signature": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        }
      }
    },
    "dto.VideoInfoRes": {
      "type": "object",
      "properties": {
        "author": {
          "description": "作者信息",
          "allOf": [
            {
              "$ref": "#/definitions/dto.UserInfoRes"
            }
          ]
        },
        "comment_count": {
          "type": "integer"
        },
        "cover_url": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "duration": {
          "type": "number"
        },
        "favorite_count": {
          "type": "integer"
        },
        "height": {
          "type": "integer"
        },
        "id": {
          "type": "string",
          "example": "0"
        },
        "is_favorite": {
          "description": "当前用户是否点赞",
          "type": "boolean"
        },
        "play_url": {
          "type": "string"
        },
        "publish_time": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "user_id": {
          "type": "string",
          "example": "0"
        },
        "view_count": {
          "type": "integer"
        },
        "width": {
          "type": "integer"
        }
      }
    },
    "res.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "msg": {
          "type": "string"
        }
      }
    }
  }
}